+++ MVCEUprava/MVCEUprava/Controllers/tblKorisnikAplikacijeController.cs
@@ -5,6 +5,7 @@ using System.Net;
 using System.Web;
 using System.Web.Mvc;
 using System.Web.Security;
+using MVCEUprava.App_Start;
 using MVCEUprava.Models;
 
 namespace MVCEUprava.Controllers
@@ -14,9 +15,13 @@ namespace MVCEUprava.Controllers
         private LicneKarteDBEntities db = new LicneKarteDBEntities();
 
         //GET: tblKorisnikAplikacije
+        [KorisnikAplikacijeAuthorize]
         public ActionResult Index()
         {
-            var tblKorisnikAplikacijes = db.tblKorisnikAplikacijes.Include(t => t.tblIzdavalac);
+            HttpCookie authCookie = System.Web.HttpContext.Current.Request.Cookies[FormsAuthentication.FormsCookieName];
+            FormsAuthenticationTicket ticket = FormsAuthentication.Decrypt(authCookie.Value);
+            string name = ticket.Name;
+            var tblKorisnikAplikacijes = db.tblKorisnikAplikacijes.Include(t => t.tblIzdavalac).Where(x=> x.Email == name).ToList();
             return View(tblKorisnikAplikacijes.ToList());
         }
 
@@ -84,27 +89,27 @@ namespace MVCEUprava.Controllers
             return View("Login", tblKorisnikAplikacije);
         }
 
-        // GET: tblKorisnikAplikacije/Details/5
-        public ActionResult Details(int? id)
-        {
-            if (id == null)
-            {
-                return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
-            }
-            tblKorisnikAplikacije tblKorisnikAplikacije = db.tblKorisnikAplikacijes.Find(id);
-            if (tblKorisnikAplikacije == null)
-            {
-                return HttpNotFound();
-            }
-            return View(tblKorisnikAplikacije);
-        }
+        //// GET: tblKorisnikAplikacije/Details/5
+        //public ActionResult Details(int? id)
+        //{
+        //    if (id == null)
+        //    {
+        //        return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
+        //    }
+        //    tblKorisnikAplikacije tblKorisnikAplikacije = db.tblKorisnikAplikacijes.Find(id);
+        //    if (tblKorisnikAplikacije == null)
+        //    {
+        //        return HttpNotFound();
+        //    }
+        //    return View(tblKorisnikAplikacije);
+        //}
 
-        // GET: tblKorisnikAplikacije/Create
-        public ActionResult Create()
-        {
-            ViewBag.Poslodavac = new SelectList(db.tblIzdavalacs, "Id", "Id");
-            return View();
-        }
+        //// GET: tblKorisnikAplikacije/Create
+        //public ActionResult Create()
+        //{
+        //    ViewBag.Poslodavac = new SelectList(db.tblIzdavalacs, "Id", "Id");
+        //    return View();
+        //}
 
         //Register
         // POST: tblKorisnikAplikacije/Create
@@ -149,6 +154,7 @@ namespace MVCEUprava.Controllers
         }
 
         // GET: tblKorisnikAplikacije/Edit/5
+        [KorisnikAplikacijeAuthorize]
         public ActionResult Edit(int? id)
         {
             if (id == null)
@@ -160,7 +166,16 @@ namespace MVCEUprava.Controllers
             {
                 return HttpNotFound();
             }
-            ViewBag.Poslodavac = new SelectList(db.tblIzdavalacs, "Id", "Id", tblKorisnikAplikacije.Poslodavac);
+            HttpCookie authCookie = System.Web.HttpContext.Current.Request.Cookies[FormsAuthentication.FormsCookieName];
+            FormsAuthenticationTicket ticket = FormsAuthentication.Decrypt(authCookie.Value);
+            string name = ticket.Name;
+            if (db.tblKorisnikAplikacijes.Where(x => x.Email == name && x.Id == tblKorisnikAplikacije.Id).ToList().Count() != 1)
+            {
+                ViewBag.Poslodavac = new SelectList(db.tblIzdavalacs, "Id", "Id", tblKorisnikAplikacije.Poslodavac);
+                return View(tblKorisnikAplikacije);
+            }
+
+            ViewBag.Poslodavac = new SelectList(db.tblIzdavalacs, "Id", "Naziv");
             return View(tblKorisnikAplikacije);
         }
 
@@ -169,19 +184,29 @@ namespace MVCEUprava.Controllers
         // more details see https://go.microsoft.com/fwlink/?LinkId=317598.
         [HttpPost]
         [ValidateAntiForgeryToken]
+        [KorisnikAplikacijeAuthorize]
         public ActionResult Edit([Bind(Include = "Id,Ime,Prezime,Pol,Jmbg,Adresa,Mesto,BrojTelefona,DatumRodjenja,Email,Password,Poslodavac")] tblKorisnikAplikacije tblKorisnikAplikacije)
         {
             if (ModelState.IsValid)
             {
+                HttpCookie authCookie = System.Web.HttpContext.Current.Request.Cookies[FormsAuthentication.FormsCookieName];
+                FormsAuthenticationTicket ticket = FormsAuthentication.Decrypt(authCookie.Value);
+                string name = ticket.Name;
+                if (db.tblKorisnikAplikacijes.Where(x => x.Email == name && x.Id == tblKorisnikAplikacije.Id).ToList().Count() != 1)
+                {
+                    ViewBag.Poslodavac = new SelectList(db.tblIzdavalacs, "Id", "Id", tblKorisnikAplikacije.Poslodavac);
+                    return View(tblKorisnikAplikacije);
+                }
                 db.Entry(tblKorisnikAplikacije).State = EntityState.Modified;
                 db.SaveChanges();
                 return RedirectToAction("Index");
             }
-            ViewBag.Poslodavac = new SelectList(db.tblIzdavalacs, "Id", "Id", tblKorisnikAplikacije.Poslodavac);
+            ViewBag.Poslodavac = new SelectList(db.tblIzdavalacs, "Id", "Naziv");
             return View(tblKorisnikAplikacije);
         }
 
         // GET: tblKorisnikAplikacije/Delete/5
+        [KorisnikAplikacijeAuthorize]
         public ActionResult Delete(int? id)
         {
             if (id == null)
@@ -189,6 +214,14 @@ namespace MVCEUprava.Controllers
                 return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
             }
             tblKorisnikAplikacije tblKorisnikAplikacije = db.tblKorisnikAplikacijes.Find(id);
+            HttpCookie authCookie = System.Web.HttpContext.Current.Request.Cookies[FormsAuthentication.FormsCookieName];
+            FormsAuthenticationTicket ticket = FormsAuthentication.Decrypt(authCookie.Value);
+            string name = ticket.Name;
+            if (db.tblKorisnikAplikacijes.Where(x => x.Email == name && x.Id == tblKorisnikAplikacije.Id).ToList().Count() != 1)
+            {
+                ViewBag.Poslodavac = new SelectList(db.tblIzdavalacs, "Id", "Naziv");
+                return View(tblKorisnikAplikacije);
+            }
             if (tblKorisnikAplikacije == null)
             {
                 return HttpNotFound();
@@ -199,15 +232,24 @@ namespace MVCEUprava.Controllers
         // POST: tblKorisnikAplikacije/Delete/5
         [HttpPost, ActionName("Delete")]
         [ValidateAntiForgeryToken]
+        [KorisnikAplikacijeAuthorize]
         public ActionResult DeleteConfirmed(int id)
         {
+            HttpCookie authCookie = System.Web.HttpContext.Current.Request.Cookies[FormsAuthentication.FormsCookieName];
+            FormsAuthenticationTicket ticket = FormsAuthentication.Decrypt(authCookie.Value);
+            string name = ticket.Name;
             tblKorisnikAplikacije tblKorisnikAplikacije = db.tblKorisnikAplikacijes.Find(id);
+            if (db.tblKorisnikAplikacijes.Where(x => x.Email == name && x.Id == id).ToList().Count() != 1)
+            {
+                ModelState.AddModelError("Message", "Ne možete izvršiti modifikaciju za drugog korisnika.");
+                return View(tblKorisnikAplikacije);
+                
+            }
+            
             db.tblKorisnikAplikacijes.Remove(tblKorisnikAplikacije);
             db.SaveChanges();
             return RedirectToAction("Index");
         }
-
-        
 
         protected override void Dispose(bool disposing)
         {
